Move MongoDbToElasticSearch
